refactor(elements): Move code to relevant sections in the codebase
