+++ packages/elements/.eslintrc.js
@@ -1,6 +1,8 @@
 module.exports = {
   root: true,
   extends: ['custom/browser', 'custom/typescript', 'custom/jest', 'custom/react'],
-  // rules: {},
+  rules: {
+    'import/no-unresolved': [2, { ignore: ['^~/'] }],
+  },
   // overrides: [],
 };
