+++ packages/elements/src/react/router/index.ts
@@ -1,0 +1,4 @@
+export { useNextRouter } from './next';
+export { Route, Router, useClerkRouter } from './react';
+
+export type { ClerkRouter, ClerkHostRouter } from './router';
