+++ packages/elements/tsconfig.lint.json
@@ -1,0 +1,5 @@
+{
+  "extends": "./tsconfig.json",
+  "include": ["src", "./*.ts"],
+  "exclude": ["node_modules"]
+}
